src/requirements.txt
Django==3.1.8
djangorestframework==3.11.2
django-autoslug ==1.9.8
django-debug-toolbar == 3.2.2
django-rest-auth == 0.9.5
django-environ == 0.8.1
django-filter==21.1
django-base64field==1.0
django-oauth-toolkit==1.5.0
django-extensions==3.1.5
django-rest-framework-social-oauth2==1.1.0
django-admin-rangefilter
django-jazzmin==2.4.8
psycopg2== 2.9.1
psycopg2-binary==2.8.4
requests==2.25.1
Pillow == 8.4.0